+++ src/test/java/com/ddudu/application/service/DeleteGoalServiceTest.java
@@ -6,11 +6,12 @@ import static org.assertj.core.api.Assertions.assertThatExceptionOfType;
 import com.ddudu.application.domain.goal.domain.Goal;
 import com.ddudu.application.domain.goal.exception.GoalErrorCode;
 import com.ddudu.application.domain.user.domain.User;
-import com.ddudu.application.port.out.DeleteGoalPort;
 import com.ddudu.application.port.out.GoalLoaderPort;
 import com.ddudu.application.port.out.SaveGoalPort;
 import com.ddudu.application.port.out.SignUpPort;
 import com.ddudu.application.port.out.UserLoaderPort;
+import com.ddudu.application.port.out.goal.DeleteGoalPort;
+import com.ddudu.application.service.goal.DeleteGoalService;
 import com.ddudu.fixture.BaseFixture;
 import com.ddudu.fixture.GoalFixture;
 import com.ddudu.fixture.UserFixture;
